move InstallmentLoader, InstallmentStore and LocalInstallmentLoader into its own file
